ServiceClient was a duplicate of ServiceReferring.  -removed HomeVIewController now just conforms to ServiceReferrring
