--- Say Their Names/Source/Controller/Base/ServiceClientViewController.swift
@@ -1,23 +1,0 @@
-//
-//  ServiceClientViewController.swift
-//  Say Their Names
-//
-//  Created by Joseph A. Wardell on 6/2/20.
-//  Copyright © 2020 Franck-Stephane Ndame Mpouli. All rights reserved.
-//
-
-import UIKit
-
-class ServiceClientViewController : UIViewController, ServiceClient  {
-    let service: Service
-    
-    required init(service: Service) {
-        self.service = service
-        super.init(nibName: nil, bundle: nil)
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
